+++ src/pal/tests/palsuite/miscellaneous/InterlockedExchangePointer/test1/CMakeLists.txt
@@ -3,7 +3,7 @@ cmake_minimum_required(VERSION 2.8.12.2)
 set(CMAKE_INCLUDE_CURRENT_DIR ON)
 
 set(SOURCES
-  InterlockedExchangePointer.c
+  InterlockedExchangePointer.cpp
 )
 
 add_executable(paltest_interlockedexchangepointer_test1
