+++ src/vm/threads.inl
@@ -22,6 +22,28 @@
 #include "appdomain.hpp"
 #include "frames.h"
 
+#ifndef DACCESS_COMPILE
+#ifdef FEATURE_IMPLICIT_TLS
+
+#ifndef __llvm__
+EXTERN_C __declspec(thread) ThreadLocalInfo gCurrentThreadInfo;
+#else // !__llvm__
+EXTERN_C __thread ThreadLocalInfo gCurrentThreadInfo;
+#endif // !__llvm__
+
+EXTERN_C inline Thread* STDCALL GetThread()
+{
+    return gCurrentThreadInfo.m_pThread;
+}
+
+EXTERN_C inline AppDomain* STDCALL GetAppDomain()
+{
+    return gCurrentThreadInfo.m_pAppDomain;
+}
+
+#endif // FEATURE_IMPLICIT_TLS
+#endif // !DACCESS_COMPILE
+
 #ifdef ENABLE_GET_THREAD_GENERIC_FULL_CHECK
 // See code:GetThreadGenericFullCheck
 inline /* static */ BOOL Thread::ShouldEnforceEEThreadNotRequiredContracts()
